Fixed bugs, Added todo.interfaces.ts
