--- src/conteiners/todo.models.ts
@@ -1,5 +1,0 @@
-export interface Todo {
-    id: string,
-    date: string,
-    text: string
-}
