+++ src/index.tsx
@@ -3,8 +3,6 @@ import * as ReactDOM from 'react-dom';
 
 import App from "./conteiners/App";
 
-import './style.css';
-
 ReactDOM.render (
     <App/>,
     document.getElementById("root")
