+++ webpack.config.js
@@ -1,34 +1,34 @@
-import path from 'path';
-import HtmlWebpackPlugin from 'html-webpack-plugin';
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-  entry: './src/index.tsx',
-  resolve: {
-    extensions: ['.ts', '.tsx'],
-  },
-  output: {
-    path: path.join(__dirname, '/dist'),
-    filename: 'bundle.min.js',
-  },
-  devServer: {
-    contentBase: path.join(__dirname, 'build'),
-    historyApiFallback: true,
-    port: 4000,
-    open: true,
-    hot: true,
-  },
-  module: {
-    rules: [
-      {
-        test: /\.ts(x?)$/,
-        exclude: /node_modules/,
-        use: 'ts-loader',
-      },
-    ],
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: './src/index.html',
-    }),
-  ],
+    entry: './src/index.tsx',
+    resolve: {
+        extensions: ['.ts', '.tsx', '.js']
+    },
+    output: {
+        path: path.join(__dirname, '/dist'),
+        filename: 'bundle.min.js'
+    },
+    devServer: {
+        contentBase: path.join(__dirname, 'build'),
+        historyApiFallback: true,
+        port: 4000,
+        open: true,
+        hot: true,
+    },
+    module: {
+        rules: [
+            {
+                test: /\.ts(x?)$/,
+                exclude: /node_modules/,
+                use: 'ts-loader'
+            }
+        ]
+    },
+    plugins: [
+        new HtmlWebpackPlugin({
+            template: './src/index.html'
+        })
+    ]
 };
